refactoring 'school service': refactored hooks
