+++ src/components/school/repo/db/index.js
@@ -1,3 +1,3 @@
-const schoolModel = require('../../../../services/school/model');
+const schoolModel = require('../../../../services/school/schools.model');
 
 module.exports = schoolModel;
