+++ src/services/edusharing/helpers.js
@@ -1,4 +1,4 @@
-const { schoolModel } = require('../school/model');
+const { schoolModel } = require('../school/schools.model');
 
 exports.getCounty = async (schoolId) => {
 	const school = await schoolModel.findById(schoolId);
