+++ src/services/fileStorage/strategies/awsS3.js
@@ -8,7 +8,7 @@ const fs = require('fs');
 const mongoose = require('mongoose');
 const { NotFound, BadRequest, GeneralError } = require('../../../errors');
 const logger = require('../../../logger');
-const { schoolModel } = require('../../school/model');
+const { schoolModel } = require('../../school/schools.model');
 const { StorageProviderModel } = require('../../storageProvider/model');
 const UserModel = require('../../user/model');
 const filePermissionHelper = require('../utils/filePermissionHelper');
