+++ src/services/school/hooks/populateInQuery.js
@@ -1,0 +1,1 @@
+module.exports = (context) => (context.params.query || {}).$populate;
