+++ src/services/school/hooks/setDefaultFileStorageType.js
@@ -1,0 +1,6 @@
+const { getDefaultFileStorageType } = require('./common');
+
+module.exports = (context) => {
+    context.data.fileStorageType = getDefaultFileStorageType();
+    return context;
+};
