+++ test/services/fileStorage/utils/providerAssignmentHelper.test.js
@@ -4,7 +4,7 @@ const { Configuration } = require('@hpi-schul-cloud/commons');
 const appPromise = require('../../../../src/app');
 const testObjects = require('../../helpers/testObjects')(appPromise);
 
-const { schoolModel } = require('../../../../src/services/school/model');
+const { schoolModel } = require('../../../../src/services/school/schools.model');
 
 const {
 	findProviderForSchool,
