+++ test/services/school/year.logic.test.js
@@ -1,7 +1,7 @@
 const chai = require('chai');
 const appPromise = require('../../../src/app');
 const SchoolYearFacade = require('../../../src/services/school/logic/year');
-const { schoolModel: School, yearModel: YearModel } = require('../../../src/services/school/model');
+const { schoolModel: School, yearModel: YearModel } = require('../../../src/services/school/schools.model');
 
 const { cleanup } = require('../helpers/testObjects')(appPromise);
 const { create: createSchool } = require('../helpers/services/schools')(appPromise);
